requirements.txt
# requirements.txt (things needed for program)
requests
python-dotenv
flask